Convert ThreadFragment to Jetpack Compose. (#192)
See https://developer.android.com/codelabs/jetpack-compose-migration#0 for background information.
Still in Step 1: very simple migration as a first step. bottom-up approach where we simply migrate the Thread screen view subtree that is shown when Thread Network is selected in Developer Utilities.

This is the last migration for Step 1. Next PR should handle scaffolding and navigation to complete the migration to Compose.