Finalizes the migration to Jetpack Compose.
Apologies, this is a huge PR. However, we only need to focus on the Compose artifacts that have been added to finalize that migration. Specifically:

- AppLayout
- AppNavigation
- AppViewModel
- MainActivity

and maybe check a few of the *Screen classes to verify they are properly designed now that all the Fragment artifacts have been removed.

There is one FIXME where I need TJ's help. See the FIXME[TJ] for details.
Essentially, what's a proper way to have a shared TopAppBar, but still allow composables
to modify its title.